support new xproto-xml schema elements, major overhaul of code gen
I've refactored the code-gen to no longer use a reader-monad to carry data
around.

The main reason we needed it was because we needed a lot of information to
properly resolve types in the XCB-XML syntax tree to haskell types.

Since this patch added more places where this type-resolution made sense,
I finally got sick of putting this reader monad places where it's not
convinient.

So now, the XCB-XML syntax tree data type is parameterized over the the
type of statements within it.  The first thing I do is convert the whole
thing over to haskell types, and then do generation.  This eleminates
most uses of the monad.  The few uses left I now plumb through the
correct infomation by hand.